pyproject.toml
[project]
name = "first-ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
